+++ canaille/app/configuration.py
@@ -173,6 +173,7 @@ def setup_config(app, config=None, test_config=True, env_file=".env", env_prefix
 def validate(config, validate_remote=False):
     validate_keypair(config.get("CANAILLE_OIDC"))
     validate_theme(config["CANAILLE"])
+    validate_admin_email(config["CANAILLE"])
 
     if not validate_remote:
         return
@@ -234,3 +235,10 @@ def validate_theme(config):
         os.path.join(ROOT, "themes", config["THEME"])
     ):
         raise ConfigurationException(f'Cannot find theme \'{config["THEME"]}\'')
+
+
+def validate_admin_email(config):
+    if config["ENABLE_PASSWORD_COMPROMISSION_CHECK"] and config["ADMIN_EMAIL"] is None:
+        raise ConfigurationException(
+            "You must set an administration email if you want to check if users' passwords are compromised."
+        )