+++ canaille/app/forms.py
@@ -1,7 +1,9 @@
 import datetime
+import hashlib
 import math
 import re
 
+import requests
 import wtforms.validators
 from flask import abort
 from flask import current_app
@@ -15,6 +17,7 @@ from canaille.app.i18n import DEFAULT_LANGUAGE_CODE
 from canaille.app.i18n import gettext as _
 from canaille.app.i18n import locale_selector
 from canaille.app.i18n import timezone_selector
+from canaille.app.mails_sending_conditions import check_if_send_mail_to_admins
 from canaille.backends import Backend
 
 from . import validate_uri
@@ -82,6 +85,39 @@ def password_strength_calculator(password):
         strength_score = strength_score * 100 // 4
 
     return strength_score
+
+
+def compromised_password_validator(form, field):
+    if not current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"]:
+        return None
+
+    hashed_password = hashlib.sha1(field.data.encode("utf-8")).hexdigest()
+    hashed_password_prefix, hashed_password_suffix = (
+        hashed_password[:5].upper(),
+        hashed_password[5:].upper(),
+    )
+
+    api_url = current_app.config["CANAILLE"]["API_URL_HIBP"] + hashed_password_prefix
+
+    try:
+        response = requests.api.get(api_url, timeout=10)
+    except Exception:
+        if not request_is_htmx():
+            current_app.logger.exception(
+                "Password compromise investigation failed on HIBP API."
+            )
+            check_if_send_mail_to_admins(form, api_url, hashed_password_suffix)
+        return None
+
+    decoded_response = response.content.decode("utf8").split("\r\n")
+
+    for each in decoded_response:
+        if hashed_password_suffix == each.split(":")[0]:
+            raise wtforms.ValidationError(
+                _(
+                    "This password appears on public compromission databases and is not secure."
+                )
+            )
 
 
 def email_validator(form, field):