+++ canaille/core/configuration.py
@@ -313,6 +313,27 @@ class CoreSettings(BaseModel):
     then 4096 will be retained.
     """
 
+    ADMIN_EMAIL: str | None = None
+    """Administration email contact.
+
+    In certain special cases (example : questioning about password
+    corruption), it is necessary to provide an administration contact
+    email.
+    """
+
+    ENABLE_PASSWORD_COMPROMISSION_CHECK: bool = False
+    """If :py:data:`True`, Canaille will check for password compromise on HIBP
+    every time a new password is register.
+
+    (https://haveibeenpwned.com/)
+    """
+
+    API_URL_HIBP: str = "https://api.pwnedpasswords.com/range/"
+    """Have i been pwned api url for compromission checks.
+
+    This url should not be modified.
+    """
+
     ENABLE_PASSWORD_EXPIRY_POLICY: bool = False
     """
     """