+++ tests/app/test_forms.py
@@ -1,4 +1,6 @@
 import datetime
+import logging
+from unittest import mock
 
 import pytest
 import wtforms
@@ -7,6 +9,7 @@ from flask import current_app
 from werkzeug.datastructures import ImmutableMultiDict
 
 from canaille.app.forms import DateTimeUTCField
+from canaille.app.forms import compromised_password_validator
 from canaille.app.forms import password_length_validator
 from canaille.app.forms import password_too_long_validator
 from canaille.app.forms import phone_number
@@ -295,14 +298,14 @@ def test_password_strength_progress_bar(testclient, logged_user):
         "/profile/user/settings",
         {
             "csrf_token": res.form["csrf_token"].value,
-            "password1": "new_password",
+            "password1": "i'm a little pea",
         },
         headers={
             "HX-Request": "true",
             "HX-Trigger-Name": "password1",
         },
     )
-    res.mustcontain('data-percent="50"')
+    res.mustcontain('data-percent="100"')
 
 
 def test_maximum_password_length_config(testclient):
@@ -333,3 +336,64 @@ def test_maximum_password_length_config(testclient):
     password_too_long_validator(None, Field("a" * 4096))
     with pytest.raises(wtforms.ValidationError):
         password_too_long_validator(None, Field("a" * 4097))
+
+
+@mock.patch("requests.api.get")
+def test_compromised_password_validator(api_get, testclient):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+
+    # This content simulates a result from the hibp api containing the suffixes of the following password hashes: 'password', '987654321', 'correct horse battery staple', 'zxcvbn123', 'azertyuiop123'
+    class Response:
+        content = b"1E4C9B93F3F0682250B6CF8331B7EE68FD8:3\r\nCAA6D483CC3887DCE9D1B8EB91408F1EA7A:3\r\nAD6438836DBE526AA231ABDE2D0EEF74D42:3\r\n8289894DDB6317178960AB5AE98B81BBF97:1\r\n5FF0B6F9EAC40D5CA7B4DAA7B64F0E6F4AA:2\r\n"
+
+    api_get.return_value = Response
+
+    class Field:
+        def __init__(self, data):
+            self.data = data
+
+    compromised_password_validator(None, Field("i'm a little pea"))
+    compromised_password_validator(None, Field("i'm a little chickpea"))
+    compromised_password_validator(None, Field("i'm singing in the rain"))
+    with pytest.raises(wtforms.ValidationError):
+        compromised_password_validator(None, Field("password"))
+    with pytest.raises(wtforms.ValidationError):
+        compromised_password_validator(None, Field("987654321"))
+    with pytest.raises(wtforms.ValidationError):
+        compromised_password_validator(None, Field("correct horse battery staple"))
+    with pytest.raises(wtforms.ValidationError):
+        compromised_password_validator(None, Field("zxcvbn123"))
+    with pytest.raises(wtforms.ValidationError):
+        compromised_password_validator(None, Field("azertyuiop123"))
+
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = False
+    assert compromised_password_validator(None, Field("password")) is None
+
+
+@mock.patch("requests.api.get")
+def test_compromised_password_validator_with_failure_of_api_request_without_form_validation(
+    api_get, testclient, logged_user, caplog
+):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+    api_get.side_effect = mock.Mock(side_effect=Exception())
+
+    res = testclient.get("/profile/user/settings")
+    res = testclient.post(
+        "/profile/user/settings",
+        {
+            "csrf_token": res.form["csrf_token"].value,
+            "password1": "correct horse battery staple",
+        },
+        headers={
+            "HX-Request": "true",
+            "HX-Trigger-Name": "password1",
+        },
+    )
+
+    res.mustcontain('data-percent="100"')
+
+    assert (
+        "canaille",
+        logging.ERROR,
+        "Password compromise investigation failed on HIBP API.",
+    ) not in caplog.record_tuples