+++ tests/core/test_profile_settings.py
@@ -156,6 +156,176 @@ def test_profile_settings_too_long_password(testclient, logged_user):
     )
 
 
+@mock.patch("requests.api.get")
+def test_profile_settings_compromised_password(api_get, testclient, logged_user):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+    """Tests if password is compromised."""
+
+    # This content simulates a result from the hibp api containing the suffixes of the following password hashes: 'password', '987654321', 'correct horse battery staple', 'zxcvbn123', 'azertyuiop123'
+    class Response:
+        content = b"1E4C9B93F3F0682250B6CF8331B7EE68FD8:3\r\nCAA6D483CC3887DCE9D1B8EB91408F1EA7A:3\r\nAD6438836DBE526AA231ABDE2D0EEF74D42:3\r\n8289894DDB6317178960AB5AE98B81BBF97:1\r\n5FF0B6F9EAC40D5CA7B4DAA7B64F0E6F4AA:2\r\n"
+
+    api_get.return_value = Response
+
+    def with_different_values(password, message):
+        res = testclient.get("/profile/user/settings")
+        res = testclient.post(
+            "/profile/user/settings",
+            {
+                "csrf_token": res.form["csrf_token"].value,
+                "password1": password,
+            },
+            headers={
+                "HX-Request": "true",
+                "HX-Trigger-Name": "password1",
+            },
+        )
+        res.mustcontain(message)
+
+    with_different_values(
+        "password",
+        "This password appears on public compromission databases and is not secure.",
+    )
+    with_different_values(
+        "azertyuiop123",
+        "This password appears on public compromission databases and is not secure.",
+    )
+    with_different_values("a" * 1000, 'data-percent="25"')
+    with_different_values("i'm a little pea", 'data-percent="100"')
+
+
+@mock.patch("requests.api.get")
+def test_profile_settings_compromised_password_request_api_failed_but_password_updated(
+    api_get, testclient, logged_user, backend, caplog
+):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+    api_get.side_effect = mock.Mock(side_effect=Exception())
+
+    current_app.config["CANAILLE"]["ACL"]["ADMIN"]["FILTER"] = {"groups": "admins"}
+
+    res = testclient.get("/profile/user/settings", status=200)
+
+    res.form["password1"] = "123456789"
+    res.form["password2"] = "123456789"
+
+    res = res.form.submit(name="action", value="edit-settings")
+
+    assert (
+        "canaille",
+        logging.ERROR,
+        "Password compromise investigation failed on HIBP API.",
+    ) in caplog.record_tuples
+    assert (
+        "error",
+        "Password compromise investigation failed. Please contact the administrators.",
+    ) in res.flashes
+    assert ("success", "Profile updated successfully.") in res.flashes
+
+    backend.reload(logged_user)
+
+    assert logged_user.user_name == "user"
+    assert backend.check_user_password(logged_user, "123456789")[0]
+
+
+@mock.patch("requests.api.get")
+def test_compromised_password_validator_with_failure_of_api_request_and_success_mail_to_admin_from_settings_form(
+    api_get, testclient, backend, user, logged_user, caplog, smtpd
+):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+    api_get.side_effect = mock.Mock(side_effect=Exception())
+
+    res = testclient.get("/profile/user/settings", status=200)
+
+    res.form.user = user
+    res.form["password1"] = "123456789"
+    res.form["password2"] = "123456789"
+
+    res = res.form.submit(name="action", value="edit-settings")
+
+    assert (
+        "canaille",
+        logging.ERROR,
+        "Password compromise investigation failed on HIBP API.",
+    ) in caplog.record_tuples
+    assert (
+        "error",
+        "Password compromise investigation failed. Please contact the administrators.",
+    ) in res.flashes
+    assert (
+        "info",
+        "We have informed your administrator about the failure of the password compromise investigation.",
+    ) in res.flashes
+    assert ("success", "Profile updated successfully.") in res.flashes
+    assert len(smtpd.messages) == 1
+
+
+@mock.patch("requests.api.get")
+def test_compromised_password_validator_with_failure_of_api_request_and_fail_to_send_mail_to_admin_from_settings_form(
+    api_get, testclient, backend, user, logged_user, caplog, smtpd
+):
+    current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+    api_get.side_effect = mock.Mock(side_effect=Exception())
+    current_app.config["CANAILLE"]["SMTP"]["TLS"] = False
+
+    res = testclient.get("/profile/user/settings", status=200)
+    res.form.user = user
+    res.form["password1"] = "123456789"
+    res.form["password2"] = "123456789"
+
+    res = res.form.submit(name="action", value="edit-settings")
+
+    assert (
+        "canaille",
+        logging.ERROR,
+        "Password compromise investigation failed on HIBP API.",
+    ) in caplog.record_tuples
+    assert (
+        "error",
+        "Password compromise investigation failed. Please contact the administrators.",
+    ) in res.flashes
+    assert (
+        "error",
+        "An error occurred while communicating the incident to the administrators. "
+        "Please update your password as soon as possible. "
+        "If this still happens, please contact the administrators.",
+    ) in res.flashes
+    assert ("success", "Profile updated successfully.") in res.flashes
+    assert len(smtpd.messages) == 0
+
+
+@mock.patch("requests.api.get")
+def test_compromised_password_validator_with_failure_of_api_request_without_smtp_or_without_admin_email_from_settings_form(
+    api_get, testclient, backend, user, logged_user, caplog
+):
+    def without_smtp_or_without_admin_email(smtp, mail):
+        current_app.config["CANAILLE"]["ENABLE_PASSWORD_COMPROMISSION_CHECK"] = True
+        api_get.side_effect = mock.Mock(side_effect=Exception())
+        current_app.config["CANAILLE"]["SMTP"] = smtp
+        current_app.config["CANAILLE"]["ADMIN_EMAIL"] = mail
+
+        res = testclient.get("/profile/user/settings", status=200)
+        res.form.user = user
+        res.form["password1"] = "123456789"
+        res.form["password2"] = "123456789"
+
+        res = res.form.submit(name="action", value="edit-settings")
+
+        assert (
+            "canaille",
+            logging.ERROR,
+            "Password compromise investigation failed on HIBP API.",
+        ) in caplog.record_tuples
+        assert (
+            "error",
+            "Password compromise investigation failed. Please contact the administrators.",
+        ) not in res.flashes
+
+    without_smtp_or_without_admin_email(
+        None, current_app.config["CANAILLE"]["ADMIN_EMAIL"]
+    )
+    without_smtp_or_without_admin_email(current_app.config["CANAILLE"]["SMTP"], None)
+
+
 def test_edition_without_groups(
     testclient,
     logged_user,
@@ -181,18 +351,18 @@ def test_edition_without_groups(
 def test_password_change(testclient, logged_user, backend, caplog):
     res = testclient.get("/profile/user/settings", status=200)
 
-    res.form["password1"] = "new_password"
-    res.form["password2"] = "new_password"
+    res.form["password1"] = "i'm a little pea"
+    res.form["password2"] = "i'm a little pea"
 
     res = res.form.submit(name="action", value="edit-settings").follow()
 
     backend.reload(logged_user)
-    assert backend.check_user_password(logged_user, "new_password")[0]
+    assert backend.check_user_password(logged_user, "i'm a little pea")[0]
 
     res = testclient.get("/profile/user/settings", status=200)
 
-    res.form["password1"] = "correct horse battery staple"
-    res.form["password2"] = "correct horse battery staple"
+    res.form["password1"] = "i'm a little chickpea"
+    res.form["password2"] = "i'm a little chickpea"
 
     res = res.form.submit(name="action", value="edit-settings")
     assert ("success", "Profile updated successfully.") in res.flashes
@@ -206,14 +376,14 @@ def test_password_change(testclient, logged_user, backend, caplog):
     res = res.follow()
 
     backend.reload(logged_user)
-    assert backend.check_user_password(logged_user, "correct horse battery staple")[0]
+    assert backend.check_user_password(logged_user, "i'm a little chickpea")[0]
 
 
 def test_password_change_fail(testclient, logged_user, backend):
     res = testclient.get("/profile/user/settings", status=200)
 
-    res.form["password1"] = "new_password"
-    res.form["password2"] = "other_password"
+    res.form["password1"] = "i'm a little pea"
+    res.form["password2"] = "i'm a little chickpea"
 
     res = res.form.submit(name="action", value="edit-settings", status=200)
 
@@ -222,7 +392,7 @@ def test_password_change_fail(testclient, logged_user, backend):
 
     res = testclient.get("/profile/user/settings", status=200)
 
-    res.form["password1"] = "new_password"
+    res.form["password1"] = "i'm a little pea"
     res.form["password2"] = ""
 
     res = res.form.submit(name="action", value="edit-settings", status=200)