+++ CHANGES.rst
@@ -1,3 +1,14 @@
+[0.0.57] - Unreleased
+---------------------
+
+Added
+^^^^^
+- Password compromission check :issue:`179`
+- :attr:`~canaille.core.configuration.CoreSettings.ADMIN_EMAIL` and
+  :attr:`~canaille.core.configuration.CoreSettings.ENABLE_PASSWORD_COMPROMISSION_CHECK` and
+  :attr:`~canaille.core.configuration.CoreSettings.API_URL_HIBP`
+  :issue:`179`
+
 [0.0.56] - 2024-11-07
 ---------------------